+++ src/BankApi.Api/Domain/Repositories/IAccountRepository.cs
@@ -12,7 +12,7 @@ namespace BankApi.Api.Domain.Repositories
 
         public Task<Account?> Withdraw(long id, decimal amount);
 
-        public Task<Dictionary<int, Account>?> Transfer(long origin, long destination, decimal amount);
+        public Task<Dictionary<long, Account>?> Transfer(long origin, long destination, decimal amount);
 
         public Task Reset();
     }
